rewrite library to provide better abstractions
